% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmc_dive.R
\name{CalcLlk}
\alias{CalcLlk}
\title{Compute log-likelihood of continuous-time Markov chain}
\usage{
CalcLlk(par, Xs, dat, len, model = "iid", model.args = NULL,
  print = FALSE)
}
\arguments{
\item{par}{parameter vector on link scale}

\item{Xs}{design matrices}

\item{dat}{data frame (see fitCTMCdive)}

\item{len}{indices that divide par into dive, surfacing parameters}

\item{model}{type of model "iid" is no temporal correlation, "cor" is temporal correlation}

\item{model.args}{contains nknots (number of knots), smooth (output of MakeSmooth function), and smoothing
parameters (lambda)}

\item{print}{TRUE if useful output to be printed}
}
\value{
log-likelihood value
}
\description{
Compute log-likelihood of continuous-time Markov chain
}
