% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmc_dive.R
\name{EDF_f}
\alias{EDF_f}
\title{EDF for smooth terms = trace(F)}
\usage{
EDF_f(X, S, lambda, Sn)
}
\arguments{
\item{X}{design matrix}

\item{S}{smoothing matrix}

\item{lambda}{smoothing parameter}

\item{Sn}{dimension of smoothing matrix}
}
\value{
trace of F = (Xt X + sp*S)^-1 Xt X
}
\description{
EDF for smooth terms = trace(F)
}
