% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmc_dive.R
\name{ExpandCovs}
\alias{ExpandCovs}
\title{Expand covariates across a time grid}
\usage{
ExpandCovs(dat, tgrid)
}
\arguments{
\item{dat}{data frame with covariates in columns and time column}

\item{tgrid}{a vector with new time grid}
}
\value{
a expanded data frame where covariates are copied from nearest time point to grid point
}
\description{
Expand covariates across a time grid
}
