% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmc_dive.R
\name{for2char}
\alias{for2char}
\title{Formula translated to a character}
\usage{
for2char(x)
}
\arguments{
\item{x}{a formula}
}
\value{
character
}
\description{
Borrowed from formula.tools
}
\author{
Christopher Brown (2018). formula.tools: Programmatic Utilities for
Manipulating Formulas, Expressions, Calls, Assignments and Other R Objects.
R package version 1.7.1.  https://CRAN.R-project.org/package=formula.tools
}
