% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmc_dive.R
\name{get_samples}
\alias{get_samples}
\title{internal function to get posterior samples}
\usage{
get_samples(mod, n = 200)
}
\arguments{
\item{mod}{a fitted model object}

\item{n}{number of samples to take}
}
\description{
internal function to get posterior samples
}
