% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmc_dive.R
\name{GetExposureEff}
\alias{GetExposureEff}
\title{Get exposure effect on predicted durations}
\usage{
GetExposureEff(mod, exp_var = "exp", base_val = 0, exp_val = NULL, nsims = 200)
}
\arguments{
\item{mod}{fitted CTMCdive model}

\item{exp_var}{name of variable(s) non-zero when exposure occurs}

\item{base_val}{what is the value to use for the baseline (default 0)}

\item{exp_val}{what value should be used for the exposure}

\item{nsims}{number of posterior simulations}
}
\value{
simulations of difference between exposure and baseline models
}
\description{
Get exposure effect on predicted durations
}
