% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmc_dive.R
\name{GetExposureEffonIntensity}
\alias{GetExposureEffonIntensity}
\title{Estimate exposure effect for dive and surface intensity}
\usage{
GetExposureEffonIntensity(mod, predgrid, basedat, nsims = 1000)
}
\arguments{
\item{mod}{fitted CTMC model}

\item{predgrid}{data frame with same columns as data but with a user-defined time grid to estimate the effect over 
must have a column which is 1 when exposure is one and zero otherwise}

\item{nsims}{number of posterior simulations, default is 1000}

\item{exp}{name of exposure indicator}
}
\value{
list with dive and surface elements which each contain the mean exposure effect and the credible interval bands
}
\description{
Estimate exposure effect for dive and surface intensity
}
