% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmc_dive.R
\name{MakeSmooth}
\alias{MakeSmooth}
\title{Compute matrices required to fit temporal smooth}
\usage{
MakeSmooth(dat, nknots = NULL)
}
\arguments{
\item{dat}{data frame (see fitCTMCdive)}

\item{nknots}{number of basis functions / knots}
}
\value{
INLA finite element matrices required to fit smooth
}
\description{
Compute matrices required to fit temporal smooth
}
