% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmc_dive.R
\name{plot.ExposureEffect}
\alias{plot.ExposureEffect}
\title{Plot exposure effect}
\usage{
\method{plot}{ExposureEffect}(expeff, pick = "all")
}
\arguments{
\item{expeff}{exposure effect estimates from GetExposureEff function}

\item{pick}{if "both" (default) then for dive and surface, otherwise for "dive" or "surface"}
}
\value{
plots of exposure effect with vertical line at point where no evidence exposure has effect from baseline
}
\description{
Plot exposure effect
}
