% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulateCTMC}
\alias{simulateCTMC}
\title{Simulate CTMCdive data}
\usage{
simulateCTMC(dive_fn, surf_fn, T, dt)
}
\arguments{
\item{dive_fn}{function that given a vector of times computes dive intensity at those times}

\item{surf_fn}{function that given a vector of times computes surface intensity at those times}

\item{dt}{time step for simulation}
}
\value{
CTMC dataset
}
\description{
Simulate CTMCdive data
}
