% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmc_dive.R
\name{simulateCTMCdive}
\alias{simulateCTMCdive}
\title{Simulate dive and surface durations for a single individual}
\usage{
simulateCTMCdive(forms, dat, par, n = 1, model = "iid",
  model.args = NULL, print = TRUE)
}
\arguments{
\item{forms}{formulae for dive and surface processes}

\item{dat}{data frame with covariates}

\item{par}{true parameter values for linear predictors with "dive" then "surface"
parameters}

\item{n}{number of records to simulate}

\item{model}{see fitCTMCdive}

\item{model.args}{seefitCTMCdive}

\item{print}{see fitCTMCdive}
}
\value{
data frame with "dive", "surface", and time columns added
}
\description{
Simulate dive and surface durations for a single individual
}
