% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulateCTMCziph}
\alias{simulateCTMCziph}
\title{Simulate CTMCdive data for Ziphius data}
\usage{
simulateCTMCziph(dive_fn, surf_fn, T, dt, kappa, sd = NULL)
}
\arguments{
\item{dive_fn}{function that given a vector of times computes dive intensity at those times}

\item{surf_fn}{function that given a vector of times computes surface intensity at those times}

\item{T}{end time}

\item{dt}{time step for simulation}

\item{kappa}{list with values for kappa for the suface and dive components (named \code{surf} and \code{dive}, respectively)}

\item{sd}{standard deviation}
}
\value{
CTMC dataset
}
\description{
Simulate CTMCdive data for Ziphius data
}
