% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmc_dive.R
\name{summary.CTMCdive}
\alias{summary.CTMCdive}
\alias{print.CTMCdive}
\title{Print summary of CTMCdive}
\usage{
\method{summary}{CTMCdive}(object, print = TRUE, ...)
}
\arguments{
\item{object}{the CTMCdive fitted model}

\item{print}{should the summary results be printed?}

\item{\dots}{unused (for S3 compatability)}
}
\value{
prints a summary
}
\description{
Print summary of CTMCdive
}
