% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmc_dive.R
\name{update.CTMCdive}
\alias{update.CTMCdive}
\title{Add or remove terms from a CTMC model and fit the new model}
\usage{
update.CTMCdive(mod, change, which = 0, print = FALSE)
}
\arguments{
\item{mod}{model to be amended}

\item{change}{change to formula to make, see ?update.formula}

\item{which}{which formula to update, 1 = dive, 2 = surface; if which = 0
then three models are fit where change is made to each formula individually
and then to both, an AIC table is printed.}
}
\value{
fitted model if which = 1,2 or list of three fitted models if which = 0
}
\description{
Add or remove terms from a CTMC model and fit the new model
}
