% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmc_dive.R
\name{AIC.CTMCdiveList}
\alias{AIC.CTMCdiveList}
\title{Akaike's An Information Criterion for lists of CTMCdive models}
\usage{
\method{AIC}{CTMCdiveList}(x, ..., k = 2)
}
\arguments{
\item{x}{a fitted detection function x}

\item{\dots}{optionally more fitted model xs.}

\item{k}{penalty per parameter to be used; the default \code{k = 2} is the "classical" AIC}
}
\description{
Calculate the AIC from a fitted model.
}
\author{
David L Miller, Richard Glennie
}
