% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmc_dive.R
\name{FitCTMCdive}
\alias{FitCTMCdive}
\title{Fits continuous-time Markov chain to dive and surface duration data}
\usage{
FitCTMCdive(forms, dat, print = TRUE)
}
\arguments{
\item{forms}{a \code{list} with formulae for \code{dive} and \code{surface} variables}

\item{dat}{a \code{data.frame} with at least three columns named \code{dive} (dive durations), \code{surface} (surface durations), and \code{time} (start time of dive); all of these must be numeric.}

\item{print}{if \code{TRUE}, useful output is printed}
}
\value{

}
\description{
Fits continuous-time Markov chain to dive and surface duration data
}
