% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmc_dive.R
\name{FitCTMCdive}
\alias{FitCTMCdive}
\title{Fits continuous-time Markov chain to dive and surface duration data}
\usage{
FitCTMCdive(
  forms,
  dat,
  print = TRUE,
  min_dwell = list(dive = 0, surface = 0),
  series = FALSE,
  re = "none",
  exp_time = NULL,
  fixed_decay = FALSE,
  knots = NULL,
  dt = NULL
)
}
\arguments{
\item{forms}{a \code{list} with formulae for \code{dive} and \code{surface} variables}

\item{dat}{a \code{data.frame} with at least three columns named \code{dive} (dive durations), \code{surface} (surface durations), and \code{time} (start time of dive); all of these must be numeric.}

\item{print}{if \code{TRUE}, useful output is printed}

\item{min_dwell}{Minimum dwell time in a state. Useful if, for example, dives are only categorised as such if they are longer than a certain interval. Named list, needs to be in the same units as \code{time}, \code{surface} and \code{dive}.}

\item{series}{if TRUE then series data, otherwise dive-by-dive data}

\item{dt}{set time-step in integration, otherwise set to total_time / 10000}

\item{rf}{if TRUE then fit a discrete random effect on dive}
}
\value{

}
\description{
Fits continuous-time Markov chain to dive and surface duration data
}
