% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmc_dive.R
\name{FitCTMCdive}
\alias{FitCTMCdive}
\title{Fits continuous-time Markov chain to dive and surface duration data}
\usage{
FitCTMCdive(forms, dat, model = "iid", model.args = NULL,
  print = TRUE, iterlim = 500)
}
\arguments{
\item{forms}{a list with formulae for "dive" and "surface" variables}

\item{dat}{a data frame with at least three columns named "dive" (dive durations), "surface" (surface durations),
and "time" (start time of dive); all of these must be numeric.}

\item{model}{"iid" fits a CTMC where durations are independent over time, "cor" fits temporal smooths so that
durations can be temporally correlated.}

\item{model.args}{if model = "cor" then model.args is a list with "nknots" set to the number of knots in the
temporal smooth; it also must contain a vector "lambda" with two elements that correspond to "dive" and "surface"
processes, if set to NA then no temporal smooth is fit, otherwise must be a positive number. This lambda is
the smoothing penalty: the larger lambda is, the smoother the estimated temporal curve.}

\item{print}{if TRUE, useful output is printed}

\item{iterlim}{iteration limit for nlm optimiser (default is 500)}
}
\value{
a CTMCdive model object: a list of the estimated results (res), variance matrix (var), fitted model returned from nlm (mod),
design matrices (Xs), formulae (forms), indices that divide par between dive and surface parameters (len),
data frame (dat), model arguments (model.args) including smoothing matrices, and the
model type (model).
}
\description{
Fits continuous-time Markov chain to dive and surface duration data
}
