% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmc_dive.R
\name{GetExposureEff}
\alias{GetExposureEff}
\title{Temporary Exposure effect extractor}
\usage{
GetExposureEff(mod, pick)
}
\arguments{
\item{mod}{model you want exposure effect for}

\item{pick}{whether to pick effect out of dive or surface model}
}
\value{
estimated exposure effect for every dive
}
\description{
Temporary Exposure effect extractor
}
