% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmc_dive.R
\name{GetExposureEff}
\alias{GetExposureEff}
\title{Estimate exposure effect for dive and surface}
\usage{
GetExposureEff(mod, predgrid, nsims = 1000)
}
\arguments{
\item{mod}{fitted CTMC model}

\item{predgrid}{data frame with same columns as data but with a user-defined time grid to estimate the effect over 
must have a column named  "exp" which is 1 during exposure period and zero otherwise}

\item{nsims}{number of posterior simulations, default is 1000}
}
\value{
list with dive and surface elements which each contain the mean exposure effect and the credible interval bands
}
\description{
Estimate exposure effect for dive and surface
}
