% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmc_dive.R
\name{MakeMatrices}
\alias{MakeMatrices}
\title{Compute matrices required to fit the model}
\usage{
MakeMatrices(forms, dat, nint = 10000)
}
\arguments{
\item{forms}{formulae}

\item{dat}{data frame (see fitCTMCdive)}

\item{nint}{number of integration points}
}
\value{
list of mesh, point A matrices, and integral A matrices
}
\description{
Compute matrices required to fit the model
}
