% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmc_dive.R
\name{MakeMatrices}
\alias{MakeMatrices}
\title{Compute matrices required to fit the model}
\usage{
MakeMatrices(forms, dat, min_dwell, series = FALSE, nint = 10000)
}
\arguments{
\item{forms}{formulae}

\item{dat}{data frame (see fitCTMCdive)}

\item{min_dwell}{start point for the integrals}

\item{series}{TRUE if data is a series rather than dive-by-dive}

\item{nint}{number of integration points}
}
\value{
list of matrices required to fit the model
}
\description{
Compute matrices required to fit the model
}
