% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmc_dive.R
\name{MakeSmooth}
\alias{MakeSmooth}
\title{Compute matrices required to fit temporal smooth}
\usage{
MakeSmooth(forms, dat, nint = 10000)
}
\arguments{
\item{forms}{formulae}

\item{dat}{data frame (see fitCTMCdive)}

\item{nint}{number of integration points}
}
\value{
list of mesh, SPDE objects, point A matrices, and integral A matrices
}
\description{
Compute matrices required to fit temporal smooth
}
