% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmc_dive.R
\name{pred_mat_maker}
\alias{pred_mat_maker}
\title{construct predictor matrix}
\usage{
pred_mat_maker(model, dat, ints)
}
\arguments{
\item{model}{fitted gam model}

\item{dat}{dataframe}

\item{time}{grid}
}
\value{
model matrix for predicting over time grid
}
\description{
construct predictor matrix
}
