% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulateCTMCziph}
\alias{simulateCTMCziph}
\title{Simulate CTMCdive data for Ziphius data}
\usage{
simulateCTMCziph(dive_fn, surf_fn, T, dt, kappa, sd = NULL, print = TRUE)
}
\arguments{
\item{dive_fn}{function that given a vector of times computes dive intensity at those times}

\item{surf_fn}{function that given a vector of times computes surface intensity at those times}

\item{dt}{time step for simulation}

\item{n}{number of dive-surfacings to simulate}
}
\value{
CTMC dataset
}
\description{
Simulate CTMCdive data for Ziphius data
}
