% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmc_dive.R
\name{summary.CTMCdive}
\alias{summary.CTMCdive}
\alias{print.CTMCdive}
\title{Print summary of CTMCdive x}
\usage{
\method{summary}{CTMCdive}(x, ...)
}
\arguments{
\item{x}{the CTMCdive fitted model x}

\item{\dots}{unused (for S3 compatability)}
}
\value{
prints a summary
}
\description{
Print summary of CTMCdive x
}
