% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmc_dive.R
\name{summary.CTMCdive}
\alias{summary.CTMCdive}
\title{Print summary of CTMCdive object}
\usage{
\method{summary}{CTMCdive}(object, ...)
}
\arguments{
\item{object}{the CTMCdive fitted model object}

\item{\dots}{unused (for S3 compatability)}
}
\value{
prints a summary
}
\description{
Print summary of CTMCdive object
}
